use lld for printing int64

+++ util.c
@@ -344,9 +344,10 @@ xmmap_commit(int64 *size) {
         }
     }
 
-    p = VirtualAlloc(NULL, (size_t)*size, MEM_COMMIT | MEM_RESERVE, PAGE_READWRITE);
+    p = VirtualAlloc(NULL, (size_t)*size, MEM_COMMIT | MEM_RESERVE,
+                     PAGE_READWRITE);
     if (p == NULL) {
-        fprintf(stderr, "Error in VirtualAlloc(%jd): %lu.\n", *size,
+        fprintf(stderr, "Error in VirtualAlloc(%lld): %lu.\n", (long long)*size,
                 GetLastError());
         fatal(EXIT_FAILURE);
     }
@@ -370,7 +371,7 @@ xmalloc(int64 size) {
         fatal(EXIT_FAILURE);
     }
     if ((p = malloc((size_t)size)) == NULL) {
-        error("Failed to allocate %zu bytes.\n", size);
+        error("Failed to allocate %lld bytes.\n", (long long)size);
         fatal(EXIT_FAILURE);
     }
     return p;
@@ -381,7 +382,7 @@ xrealloc(void *old, const int64 size) {
     void *p;
     uint64 old_save = (uint64)old;
     if (size <= 0) {
-        error("Error in xmalloc: invalid size = %ld.\n", size);
+        error("Error in xmalloc: invalid size = %lld.\n", (long long)size);
         fatal(EXIT_FAILURE);
     }
     if ((p = realloc(old, (usize)size)) == NULL) {
